requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
transformers>=4.30.0
datasets>=2.12.0
librosa>=0.10.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
wandb>=0.15.0
hydra-core>=1.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
soundfile>=0.12.0
jiwer>=3.0.0 